Fix all MSVC compiler warnings across the codebase
This commit addresses all C4244, C4267, C4189, C4456, C4457, C4459, C4100, C4505, C4005, and C4702 warnings reported by the MSVC compiler.

Changes include:
- chain.cpp: Fix uint64_t to uint32_t conversion in FilterStore::put call
- mempool.cpp: Suppress unused variable warning for reserved RBF feature
- p2p.cpp: Fix type conversions (uint64_t/size_t to uint32_t), variable shadowing (h, i, it, z, lk), and add MSVC support for [[maybe_unused]]
- stratum_server.cpp: Fix lock guard variable shadowing
- wallet_store.cpp: Suppress unused passphrase parameter when encryption is disabled
- miqminer_rpc.cpp: Fix unreferenced parameters, type conversions (size_t to int, chrono to int), and mark unused function
- main.cpp: Fix macro redefinition, unreferenced parameters/functions, and unreachable code in constexpr templates
- miqwallet.cpp: Fix int to char conversions, variable shadowing (seed), and mark 30+ unreferenced static functions with [[maybe_unused]]
- p2p_light.cpp: Fix variable shadowing in nested loop
- spv_simple.cpp: Rename local MAX_TX_SIZE to avoid global declaration hiding

All warnings have been professionally addressed while maintaining code functionality and future extensibility.

+++ src/chain.cpp
@@ -1855,7 +1855,7 @@ bool Chain::submit_block(const Block& b, std::string& err){
     {
         std::vector<uint8_t> fbytes;
         if (miq::gcs::build_block_filter(b, fbytes)) {
-            if (!g_filter_store.put(new_height, new_hash, fbytes)) {
+            if (!g_filter_store.put(static_cast<uint32_t>(new_height), new_hash, fbytes)) {
                 log_warn("FilterStore: put failed for height=" + std::to_string(new_height) +
                          " hash=" + hexstr(new_hash));
             }
